backend/requirements.txt
aniso8601==9.0.1
anyio==3.7.1
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
build==1.0.0
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.4
cryptography==41.0.2
dnspython==2.4.0
eventlet==0.33.3
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-RESTful==0.3.10
flask-restful-swagger-2==0.35
Flask-Session==0.5.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2==2.9.7
pycparser==2.21
pyproject_hooks==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.1
safrs==3.1.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
tomli==2.0.1
typing_extensions==4.7.1
uWSGI==2.0.21
Werkzeug==2.3.6
zipp==3.16.0
zope.event==5.0
zope.interface==6.0

Pillow~=10.0.1
boto3~=1.28.62
requests~=2.31.0